# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_imports_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_imports_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportCommitOut": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/aicam__modules__imports__schemas__Counts"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "counts"
        ],
        "title": "ImportCommitOut",
        "type": "object"
      },
      "ImportItem": {
        "properties": {
          "committed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "counts": {
            "$ref": "#/components/schemas/aicam__modules__imports__schemas__Counts"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "file_name",
          "counts",
          "created_by",
          "created_at",
          "committed_at"
        ],
        "title": "ImportItem",
        "type": "object"
      },
      "ImportPreviewOut": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/aicam__modules__imports__schemas__Counts"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/RowErrorOut"
            },
            "title": "Errors",
            "type": "array"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sample": {
            "items": {
              "$ref": "#/components/schemas/SampleRow"
            },
            "title": "Sample",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "file_name",
          "counts",
          "errors",
          "sample",
          "expires_at"
        ],
        "title": "ImportPreviewOut",
        "type": "object"
      },
...
      "PackageDetail": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "platform_logistics_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Logistics Status"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionDetail"
            },
            "title": "Sessions",
            "type": "array"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            },
            "title": "Timeline",
            "type": "array"
          },
          "tracking_number": {
            "title": "Tracking Number",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          },
          "warehouse_status": {
            "title": "Warehouse Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tracking_number",
          "warehouse_status",
          "platform_logistics_status",
          "verified",
          "order",
          "sessions",
          "timeline"
        ],
        "title": "PackageDetail",
        "type": "object"
      },
      "PackageItem": {
        "properties": {
          "has_clip": {
            "title": "Has Clip",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LastSession"
              },
              {
                "type": "null"
              }
            ]
          },
          "platform_order_sn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Order Sn"
          },
          "platform_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Status"
          },
          "source": {
            "anyOf": [
              {
                "enum": [
                  "API",
                  "CSV"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tracking_number": {
            "title": "Tracking Number",
            "type": "string"
          },
          "warehouse_status": {
            "title": "Warehouse Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tracking_number",
          "platform_order_sn",
          "warehouse_status",
          "platform_status",
          "source",
          "last_session",
          "has_clip"
        ],
        "title": "PackageItem",
        "type": "object"
      },
...
      "Page_ImportItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImportItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[ImportItem]",
        "type": "object"
      },
      "Page_PackageItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PackageItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[PackageItem]",
        "type": "object"
      },
...
      "QueuedOut": {
        "properties": {
          "queued": {
            "default": true,
            "title": "Queued",
            "type": "boolean"
          }
        },
        "title": "QueuedOut",
        "type": "object"
      },
...
      "RowErrorOut": {
        "properties": {
          "column": {
            "title": "Column",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "row": {
            "title": "Row",
            "type": "integer"
          }
        },
        "required": [
          "row",
          "column",
          "message"
        ],
        "title": "RowErrorOut",
        "type": "object"
      },
      "SampleRow": {
        "properties": {
          "action": {
            "enum": [
              "NEW",
              "UPDATE",
              "SKIP"
            ],
            "title": "Action",
            "type": "string"
          },
          "platform_order_sn": {
            "title": "Platform Order Sn",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "row": {
            "title": "Row",
            "type": "integer"
          },
          "tracking_number": {
            "title": "Tracking Number",
            "type": "string"
          },
          "variation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variation"
          }
        },
        "required": [
          "row",
          "tracking_number",
          "platform_order_sn",
          "product_name",
          "variation",
          "quantity",
          "action"
        ],
        "title": "SampleRow",
        "type": "object"
      },
...
      "UserBrief": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "UserBrief",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "aicam__modules__imports__schemas__Counts": {
        "properties": {
          "error": {
            "default": 0,
            "title": "Error",
            "type": "integer"
          },
          "new": {
            "default": 0,
            "title": "New",
            "type": "integer"
          },
          "skipped": {
            "default": 0,
            "title": "Skipped",
            "type": "integer"
          },
          "updated": {
            "default": 0,
            "title": "Updated",
            "type": "integer"
          }
        },
        "title": "Counts",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Hệ thống X API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/imports": {
      "get": {
        "description": "API-52: lịch sử nhập, mới nhất trước.",
        "operationId": "history_api_v1_imports_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ImportItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "imports"
        ]
      },
      "post": {
        "description": "API-50: multipart field `file` (.csv UTF-8 / .xlsx, ≤ 5 MB, ≤ 5.000 dòng) → bản xem trước 30 phút.",
        "operationId": "upload_api_v1_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "imports"
        ]
      }
    },
...
    "/api/v1/imports/{import_id}/commit": {
      "post": {
        "description": "API-51: chỉ người tạo; từ chối khi còn dòng lỗi / quá 30 phút (EX-P11).",
        "operationId": "commit_api_v1_imports__import_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportCommitOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/file": {
      "get": {
        "description": "API-54: tải file gốc (giữ 90 ngày, sau đó 410 FILE_EXPIRED).",
        "operationId": "original_file_api_v1_imports__import_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Original File",
        "tags": [
          "imports"
        ]
      }
    },
...
    "/api/v1/packages": {
      "get": {
        "description": "API-30: tra cứu kiện (FR-07.01, 07.03).",
        "operationId": "search_packages_api_v1_packages_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          },
          {
            "in": "query",
            "name": "warehouse_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "NEW",
                    "PACKING",
                    "PACKED",
                    "HANDED_OVER",
                    "DELIVERED",
                    "CANCELLED",
                    "CANCELLED_AFTER_PACK"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Status"
            }
          },
          {
            "in": "query",
            "name": "session_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "OPEN",
                    "MISMATCH",
                    "WAITING_APPROVAL",
                    "COMPLETED",
                    "CANCELLED",
                    "ABANDONED",
                    "SUPERSEDED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Status"
            }
          },
          {
            "in": "query",
            "name": "session_flag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "UNVERIFIED",
                    "CAM2_UNVERIFIED",
                    "LABEL_ON_TRAY",
                    "VIDEO_INCOMPLETE",
                    "REPACK",
                    "HAD_MISMATCH",
                    "CLOSED_BY_SUPERVISOR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Flag"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "API",
                    "CSV"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PackageItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Packages",
        "tags": [
          "packages"
        ]
      }
    },
    "/api/v1/packages/{package_id}": {
      "get": {
        "description": "API-31: chi tiết kiện, phiên, clip, dòng thời gian (FR-07.02).",
        "operationId": "package_detail_api_v1_packages__package_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Package Detail",
        "tags": [
          "packages"
        ]
      }
    },
...
    "/api/v1/shops/shopee/callback": {
      "get": {
        "description": "API-72 (công khai + `state`): đổi code → token, lưu shop, J-04 ngay; 302 về D7 `?result=`.",
        "operationId": "callback_api_v1_shops_shopee_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "shop_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback",
        "tags": [
          "shops"
        ]
      }
    },
    "/api/v1/shops/{shop_id}/sync": {
      "post": {
        "description": "API-73: đồng bộ ngay (J-04 queue `sync`); đang chạy → 409 SYNC_IN_PROGRESS.",
        "operationId": "sync_api_v1_shops__shop_id__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync",
        "tags": [
          "shops"
        ]
      }
    },
...
  }
}